refactor: change the location of the function for validation
